backend/requirements.txt
django>=4.2
gunicorn
pymysql
cryptography
djangorestframework
pandas
openpyxl
qrcode[pil]
reportlab
